Refactor app state into useQuiz hook
