+++ src/App.tsx
@@ -1,6 +1,6 @@
 // src/App.tsx
-import React, { useState, useEffect } from 'react'
-import useQuestions from './hooks/useQuestions'
+import React from 'react'
+import useQuiz from './hooks/useQuiz'
 import NavBar from './components/Layouts/NavBar'
 import MenuScreen from './components/Layouts/MenuScreen'
 import SetupScreen from './components/Settings/SetupScreen'
@@ -8,116 +8,28 @@ import ThemeScreen from './components/Settings/ThemeScreen'
 import StartScreen from './components/Quiz/StartScreen'
 import QuizScreen from './components/Quiz/QuizScreen'
 import ResultScreen from './components/Quiz/ResultScreen'
-import type { OpenTDBQuestion, Settings } from './types'
+import type { OpenTDBQuestion } from './types'
 
 /**
  * App: Root component controlling quiz flow and in-game menu drawer.
  */
 export default function App() {
-  // 1. Store user settings: theme, category, amount, difficulty
-  const defaultSettings: Settings = {
-    theme: 'default',
-    category: 0,
-    amount: 10,
-    difficulty: 'any',
-  }
-  const [settings, setSettings] = useState<Settings>(() => {
-    const saved = localStorage.getItem('settings')
-    if (saved) {
-      try {
-        return { ...defaultSettings, ...JSON.parse(saved) }
-      } catch {
-        // fall through to defaults
-      }
-    }
-    return defaultSettings
-  })
-
-  // 2. Fetch questions whenever gameplay settings change
-  const { questions, loading, error } = useQuestions(
-    settings.amount,
-    settings.category,
-    settings.difficulty
-  )
-
-  // 3. UI stage: controls which panel/screen is visible
-  const [stage, setStage] = useState<
-    'menu' | 'settings' | 'theme' | 'start' | 'quiz' | 'result'
-  >('start')
-
-  // 4. Record of user answers: question ID -> correct?
-  const [responses, setResponses] = useState<Record<string, boolean>>({})
-
-  /** Toggle the in-game menu drawer */
-  const toggleMenu = () => {
-    setStage(prev => (prev === 'menu' ? 'start' : 'menu'))
-  }
-
-  // Close menu when Escape key is pressed
-  useEffect(() => {
-    const handleKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape' && stage === 'menu') {
-        toggleMenu()
-      }
-    }
-    window.addEventListener('keydown', handleKey)
-    return () => window.removeEventListener('keydown', handleKey)
-  }, [stage])
-
-  /** Apply a new theme and close menu */
-  const handleThemeSelect = (theme: Settings['theme']) => {
-    setSettings(prev => ({ ...prev, theme }))
-  }
-
-  /** Save gameplay options from SetupScreen and close menu */
-  const handleSaveSettings = ({ category, amount, difficulty }: Omit<Settings, 'theme'>) => {
-    setSettings(prev => ({ ...prev, category, amount, difficulty }))
-    setStage('start')
-  }
-
-  /** Close menu without saving */
-  const handleCancel = () => {
-    setStage('start')
-  }
-
-  /** Start the quiz */
-  const handleStart = () => setStage('quiz')
-
-  /** Persist settings whenever they change */
-  useEffect(() => {
-    localStorage.setItem('settings', JSON.stringify(settings))
-  }, [settings])
-
-  /** Record a user's answer, ignoring duplicates */
-  const handleAnswer = (id: string, correct: boolean) => {
-    setResponses(prev => (prev[id] != null ? prev : { ...prev, [id]: correct }))
-  }
-
-  /** Auto-advance to results when all questions are answered */
-  useEffect(() => {
-    if (
-      stage === 'quiz' &&
-      questions.length > 0 &&
-      Object.keys(responses).length === questions.length
-    ) {
-      setStage('result')
-    }
-  }, [stage, questions, responses])
-
-  /** Restart quiz from the beginning */
-  const handleRestart = () => {
-    setResponses({})
-    setStage('start')
-  }
-
-  /** Reset progress when gameplay settings change */
-  useEffect(() => {
-    setResponses({})
-    setStage('start')
-  }, [settings.category, settings.amount, settings.difficulty])
-
-  // Compute current score: count of correct answers
-  const score = Object.values(responses).filter(Boolean).length
+  const {
+    settings,
+    stage,
+    questions,
+    loading,
+    error,
+    score,
+    toggleMenu,
+    selectTheme,
+    saveSettings,
+    cancel,
+    startQuiz,
+    answerQuestion,
+    restartQuiz,
+    completeQuiz,
+  } = useQuiz()
 
   // Render loading or error states
   if (loading) return <div className="text-center mt-8">Loading questions…</div>
@@ -140,9 +52,9 @@ export default function App() {
           <div className="inset-x-0 top-0 md:inset-y-0 md:right-0 z-20">
             <MenuScreen
               initialSettings={settings}
-              onSaveSettings={handleSaveSettings}
-              onThemeSelect={handleThemeSelect}
-              onCancel={handleCancel}
+              onSaveSettings={saveSettings}
+              onThemeSelect={selectTheme}
+              onCancel={cancel}
             />
           </div>
         </>
@@ -153,26 +65,26 @@ export default function App() {
         {stage === 'settings' && (
           <SetupScreen
             initial={settings}
-            onSave={handleSaveSettings}
-            onCancel={handleCancel}
+            onSave={saveSettings}
+            onCancel={cancel}
           />
         )}
 
         {stage === 'theme' && (
           <ThemeScreen
             initialTheme={settings.theme}
-            onThemeSelect={handleThemeSelect}
-            onCancel={handleCancel}
+            onThemeSelect={selectTheme}
+            onCancel={cancel}
           />
         )}
 
-        {stage === 'start' && <StartScreen onStart={handleStart} />}
+        {stage === 'start' && <StartScreen onStart={startQuiz} />}
 
         {stage === 'quiz' && (
           <QuizScreen
             questions={questions as OpenTDBQuestion[]}
-            onAnswered={handleAnswer}
-            onComplete={() => setStage('result')}
+            onAnswered={answerQuestion}
+            onComplete={completeQuiz}
           />
         )}
 
@@ -180,7 +92,7 @@ export default function App() {
           <ResultScreen
             score={score}
             total={questions.length}
-            onRestart={handleRestart}
+            onRestart={restartQuiz}
           />
         )}
       </>
